ALTAPPS-335: Move HyperskillUrlPath and HyperskillUrlBuilder to core.domain.url package
