+++ shared/src/commonMain/kotlin/org/hyperskill/app/core/domain/url/HyperskillUrlBuilder.kt
@@ -1,10 +1,9 @@
-package org.hyperskill.app.open_url_in_web.domain.builder
+package org.hyperskill.app.core.domain.url
 
 import io.ktor.http.URLBuilder
 import io.ktor.http.URLProtocol
 import io.ktor.http.Url
 import org.hyperskill.app.config.BuildKonfig
-import org.hyperskill.app.open_url_in_web.domain.model.HyperskillUrlPath
 
 object HyperskillUrlBuilder {
     fun build(path: HyperskillUrlPath): Url {
